requirements.txt
# Author    : Clinton Lohr
# Date      : 04/26/2023
# Course    : CS 493 - Cloud Application Development
# Assignment: 04 - Intermediate REST API


Flask==2.3.2
google-cloud-datastore==2.15.0
python-jose
flask-cors
six~=1.16.0
python-dotenv
requests
authlib~=1.2.0
protobuf==3.20.*

google~=3.0.0